// src/models/IoK8sApiCertificatesV1alpha1ClusterTrustBundle/types.ts
import { IoK8sApiCertificatesV1alpha1ClusterTrustBundleSpecModelOptions as IoK8sApiCertificatesV1alpha1ClusterTrustBundleSpec } from '../IoK8sApiCertificatesV1alpha1ClusterTrustBundleSpec';
import { IoK8sApimachineryPkgApisMetaV1ObjectMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ObjectMeta } from '../IoK8sApimachineryPkgApisMetaV1ObjectMeta';

export interface IoK8sApiCertificatesV1alpha1ClusterTrustBundleModelOptions {
  apiVersion?: string;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ObjectMeta;
  spec: IoK8sApiCertificatesV1alpha1ClusterTrustBundleSpec;
}