// src/models/IoK8sApiCertificatesV1alpha1ClusterTrustBundleList/types.ts
import { IoK8sApiCertificatesV1alpha1ClusterTrustBundleModelOptions as IoK8sApiCertificatesV1alpha1ClusterTrustBundle } from '../IoK8sApiCertificatesV1alpha1ClusterTrustBundle';
import { IoK8sApimachineryPkgApisMetaV1ListMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ListMeta } from '../IoK8sApimachineryPkgApisMetaV1ListMeta';

export interface IoK8sApiCertificatesV1alpha1ClusterTrustBundleListModelOptions {
  apiVersion?: string;
  items: Array<IoK8sApiCertificatesV1alpha1ClusterTrustBundle>;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ListMeta;
}